requirements.txt
ics==0.6
pytz==2019.3
Flask==1.1.1
flask-wtf==0.14.2
flask-mail==0.9.1
flask_breadcrumbs==0.4.0
flask_bootstrap4==4.0.2
flask_migrate==2.2.1
flask_sqlalchemy==2.3.2
flask_socketio==4.2.1
flask_cors==3.0.8
spotipy==2.4.4
wget==3.2
gevent==1.4.0
python-docx==0.8.10